+++ src/utils/log/index.js
@@ -1,0 +1,11 @@
+
+export default {
+
+  warn: message => {
+    console.warn(`Redux-Honey: \n ${message}`);
+  },
+
+  error: message => {
+    console.error(`Redux-Honey: \n ${message}`);
+  }
+}