+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 
 # ChangeLog
 
-### **5.0.1**
-- Added improved error message handing
+### **5.1.0**
+- Added a method ```state.resetKey()``` that allows the resetting of a single key vs resetting the entire state with ```state.reset()```
+- Improved error messages for potential points of failure
 
 ### **5.0.1**
 - Removed spread operator use in ```extract``` method